requirements.txt
black
flake8
isort
coverage
pylint_django
pytest
pytest-django
pytest-cov
pytest-mock
python-dotenv
pylint
Django
djangorestframework
gunicorn
psycopg[binary]